+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2021-01-15
+
+### Added
+
+- Fully working api endpoint to store an invoice in the database.
+- Fully working api endpoint to get an invoice from the database by its invoice number.
+- Fully working api endpoint to update an invoice in the database.
+- Fully working api endpoint to delete an invoice in the database.
+- Integration tests for the: create-invoice, get-by-invoice-number, update-invoice and delete-invoice endpoints. 
+- Documentation for each of these enpoints.
+
+## [1.4.0] - 2021-01-15
+
+### Added
+
+- Delete-invoice api endpoint.
+- Integration tests for the delete-invoice api endpoint.
+- Integration test for trying to update an unexisting invoice.
+- Documentation for the delete-invoice api endpoint.
+
 ## [1.3.1] - 2021-01-15
 
 ### Fixed